Feature:  Loot Table

+++ code/player.py
@@ -321,22 +321,26 @@ class Player(Entity):
             if particle.hitbox.colliderect(self.hitbox):
                 if 'hurt' not in self.state and not self.invulnerable and particle.affects_player:
                     # Shield test
-                    if ('attacking' not in self.state
-                        and ((self.direction_label == 'up' and particle.direction_vector.y > 0)
-                             or (self.direction_label == 'down' and particle.direction_vector.y < 0)
-                             or (self.direction_label == 'left' and particle.direction_vector.x > 0)
-                             or (self.direction_label == 'right' and particle.direction_vector.x < 0))):
+                    if (not particle.bypasses_shield
+                            and 'attacking' not in self.state
+                            and ((self.direction_label == 'up' and particle.direction_vector.y > 0)
+                                 or (self.direction_label == 'down' and particle.direction_vector.y < 0)
+                                 or (self.direction_label == 'left' and particle.direction_vector.x > 0)
+                                 or (self.direction_label == 'right' and particle.direction_vector.x < 0))):
                         # Successful block ! Should add a sound
                         particle.kill()
                     else:
-                        self.state = 'hurt_p'
-                        self.hurt_starting_time = pygame.time.get_ticks()
-                        self.hurt_animation_starting_time = self.hurt_starting_time
-                        self.invulnerable = True
-                        self.direction_vector = particle.direction_vector
-                        self.hitbox.x += self.current_speed*self.direction_vector.x
-                        self.hitbox.y += self.current_speed*self.direction_vector.y
-                        self.health -= particle.collision_damage
+                        if particle.collision_damage != 0:
+                            self.state = 'hurt_p'
+                            self.hurt_starting_time = pygame.time.get_ticks()
+                            self.hurt_animation_starting_time = self.hurt_starting_time
+                            self.invulnerable = True
+                            self.direction_vector = particle.direction_vector
+                            self.hitbox.x += self.current_speed*self.direction_vector.x
+                            self.hitbox.y += self.current_speed*self.direction_vector.y
+                            self.health -= particle.collision_damage
+                        else:
+                            particle.effect()
                         particle.kill()
 
     def move(self):
